refactor: isolate tutorial feature components into src/features/tutorial
