+++ memory-bank/systemPatterns.md
@@ -77,7 +77,7 @@ tags: []
   - Implements Natural Language Semantic Filtering (`useAiSearch` and `useDebouncedSemanticSearch` hooks) that uses local LiteRT-LM Gemma model to parse user's query into structured database filters (rarity, category, color, keyword). The implementation is fully modularized by extracting components (`CardsGrid.tsx`, `EmptyState.tsx`, `ExtractedFiltersDisplay.tsx`, `AiWarningModal.tsx`) to comply with strict ESLint line count rules.
 - **Tutorial Spotlight & Position Synchronization**:
   - Tutorial spotlight positioning, window resize/scroll event listeners, click-blocking logic, and database mock insertions are fully encapsulated in the `useSpotlight` custom hook (`src/hooks/useSpotlight.ts`).
-  - Decouples DOM calculations and window event handlers from the `InteractiveTutorial` overlay component (`src/components/organisms/InteractiveTutorial.tsx`), maintaining clean separation of concerns and keeping component file size well within constraints (under 300 lines).
+  - Decouples DOM calculations and window event handlers from the `InteractiveTutorial` overlay component (`src/features/tutorial/components/InteractiveTutorial.tsx`), maintaining clean separation of concerns and keeping component file size well within constraints (under 300 lines).
 - **Chrome Extension API & Prompt Injection Decoupling**:
   - Chrome extension connection monitoring (ping/retry) and prompt injection logic (sending to Midjourney tab, updating usage stats, slot value persistence) are fully decoupled from `Workbench.tsx` and encapsulated into reusable hooks `useChromeTabConnection.ts` and `usePromptInjector.ts`.
   - The Workbench UI and form layouts are modularized into focused components (`WorkbenchView.tsx`, `Cauldron.tsx`, `RecipeEditor.tsx`, `RecipeForm.tsx`) and hooks (`useWorkbenchCore`), keeping each component file size under the 300-line limit and functions under the 50-line limit, and enabling straightforward unit testing without mocking heavy layout structures.
