+++ src/features/tutorial/components/OnboardingGuide/BubblesIllustration.tsx
