+++ src/features/tutorial/components/OnboardingGuide/ClickCardIllustration.tsx
