+++ src/features/tutorial/components/OnboardingGuide/MoveIllustration.tsx
