+++ src/features/tutorial/components/OnboardingGuide/OnboardingSteps.tsx
