+++ src/features/tutorial/components/OnboardingGuide/RarityIllustration.tsx
