+++ src/features/tutorial/components/OnboardingGuide/WorkbenchEditorIllustration.tsx
