+++ tests/unit/components/organisms/OnboardingGuide.test.tsx
@@ -1,5 +1,5 @@
-import { OnboardingGuide } from "@/components/organisms/OnboardingGuide"
 import { LanguageProvider } from "@/contexts/LanguageContext"
+import { OnboardingGuide } from "@/features/tutorial/components/OnboardingGuide"
 import { fireEvent, render, screen } from "@testing-library/react"
 import { describe, expect, it, vi } from "vitest"
